+++ src/pages/BudgetGlossaryPage.tsx
@@ -1,0 +1,5 @@
+import GlossaryPage from './GlossaryPage.tsx';
+
+export default function BudgetGlossaryPage() {
+  return <GlossaryPage domain="budget" />;
+}