+++ src/pages/RBIGlossaryPage.tsx
@@ -1,0 +1,5 @@
+import GlossaryPage from './GlossaryPage.tsx';
+
+export default function RBIGlossaryPage() {
+  return <GlossaryPage domain="rbi" />;
+}